Fixing filtered select bug
If you applied a filter to the table it would snapshot the rows state, so when you would try to select or deselect the row it would only work once. It would throw it out as going to the same state. So I changed it so it would check the the rows in state to see if the id exists in the object. This may not work for deep rows or be the perfect solution, but this is the idea behind the solution.